editing of text annotations now working

+++ web/apps/web_copo/views.py
@@ -54,15 +54,21 @@ def login(request):
 
 
 def test(request):
+    html = ""
     records = Annotation().get_all_records("5d6d2df56f3ca55f1ef968dd")
-    folder_name = str(uuid.uuid1())
-    full_path = os.path.join('/home/fshaw/Desktop/htmls', folder_name)
-    os.makedirs(full_path)
-    run("ebook-convert /home/fshaw/pdf/1.pdf " + full_path + " --no-images --pretty-print --insert-blank-line")
-    with open(os.path.join(full_path, "index.html"), 'r') as f:
-        html = f.read()
-    shutil.rmtree(full_path)
+    if "annotation_html" not in request.session:
 
+        folder_name = str(uuid.uuid1())
+        full_path = os.path.join('/home/fshaw/Desktop/htmls', folder_name)
+        os.makedirs(full_path)
+        run("ebook-convert /home/fshaw/pdf/1.pdf " + full_path + " --no-images --pretty-print --insert-blank-line")
+        with open(os.path.join(full_path, "index.html"), 'r') as f:
+           html = f.read()
+        shutil.rmtree(full_path)
+        request.session["annotation_html"] = html
+    else:
+        print("using session text data")
+        html = request.session["annotation_html"]
     return render(request, 'copo/test_page.html',
                   {"records": records, "file_id": "5d6d2df56f3ca55f1ef968dd", "file_name": "2.pdf", "html": html})
